ref(slack): Move issue / incident message building to a module
