+++ src/sentry/integrations/slack/unfurl/incidents.py
@@ -1,12 +1,14 @@
 import re
 from typing import List
+
 from django.db.models import Q
+from django.http.request import HttpRequest
 
 from sentry.incidents.models import Incident
-from sentry.integrations.slack.utils import build_incident_attachment
+from sentry.integrations.slack.message_builder.incidents import build_incident_attachment
+from sentry.models.integration import Integration
 
 from . import Handler, UnfurlableUrl, UnfurledUrl, make_type_coercer
-
 
 map_incident_args = make_type_coercer(
     {
@@ -16,7 +18,9 @@ map_incident_args = make_type_coercer(
 )
 
 
-def unfurl_incidents(request, integration, links: List[UnfurlableUrl]) -> UnfurledUrl:
+def unfurl_incidents(
+    request: HttpRequest, integration: Integration, links: List[UnfurlableUrl]
+) -> UnfurledUrl:
     filter_query = Q()
     # Since we don't have real ids here, we use the org slug so that we can
     # make sure the identifiers correspond to the correct organization.
