+++ src/sentry/integrations/slack/utils.py
@@ -1,45 +1,25 @@
 import logging
 import time
+from typing import List, Tuple
+from urllib.parse import parse_qs, urlencode, urlparse
 
-from django.core.cache import cache
 from django.core.exceptions import ValidationError
 from django.http import Http404
-from urllib.parse import urlparse, urlencode, parse_qs
 
-from sentry import tagstore
 from sentry.constants import ObjectStatus
-from sentry.utils import json
-from sentry.utils.assets import get_asset_url
-from sentry.utils.dates import to_timestamp
-from sentry.utils.http import absolute_uri
-from sentry.models import (
-    ActorTuple,
-    GroupStatus,
-    GroupAssignee,
-    OrganizationMember,
-    Project,
-    User,
-    Identity,
-    IdentityProvider,
-    Integration,
-    Organization,
-    Team,
-    ReleaseProject,
-)
+from sentry.models import IdentityProvider, Integration, Organization
 from sentry.shared_integrations.exceptions import (
     ApiError,
     DuplicateDisplayNameError,
     IntegrationError,
 )
-from sentry.integrations.metric_alerts import incident_attachment_info
+from sentry.utils import json
 
 from .client import SlackClient
 
 logger = logging.getLogger("sentry.integrations.slack")
 
 # Attachment colors used for issues with no actions take
-ACTIONED_ISSUE_COLOR = "#EDEEEF"
-RESOLVED_COLOR = "#4dc771"
 LEVEL_TO_COLOR = {
     "debug": "#fbe14f",
     "info": "#2788ce",
@@ -47,317 +27,39 @@ LEVEL_TO_COLOR = {
     "error": "#E03E2F",
     "fatal": "#FA4747",
 }
+
+ACTIONED_ISSUE_COLOR = "#EDEEEF"
+INCIDENT_RESOLVED_COLOR = "#4dc771"
+
+
 MEMBER_PREFIX = "@"
 CHANNEL_PREFIX = "#"
 strip_channel_chars = "".join([MEMBER_PREFIX, CHANNEL_PREFIX])
 SLACK_DEFAULT_TIMEOUT = 10
 
 
-def get_integration_type(integration):
+def get_integration_type(integration: Integration):
     metadata = integration.metadata
     # classic bots had a user_access_token in the metadata
     default_installation = "classic_bot" if "user_access_token" in metadata else "workspace_app"
     return metadata.get("installation_type", default_installation)
 
 
-def format_actor_option(actor):
-    if isinstance(actor, User):
-        return {"text": actor.get_display_name(), "value": f"user:{actor.id}"}
-    if isinstance(actor, Team):
-        return {"text": f"#{actor.slug}", "value": f"team:{actor.id}"}
-
-    raise NotImplementedError
-
-
-def get_member_assignees(group):
-    queryset = (
-        OrganizationMember.objects.filter(
-            user__is_active=True,
-            organization=group.organization,
-            teams__in=group.project.teams.all(),
-        )
-        .distinct()
-        .select_related("user")
-    )
-
-    members = sorted(queryset, key=lambda u: u.user.get_display_name())
-
-    return [format_actor_option(u.user) for u in members]
-
-
-def get_team_assignees(group):
-    return [format_actor_option(u) for u in group.project.teams.all()]
-
-
-def get_assignee(group):
-    try:
-        assigned_actor = GroupAssignee.objects.get(group=group).assigned_actor()
-    except GroupAssignee.DoesNotExist:
-        return None
-
-    try:
-        return format_actor_option(assigned_actor.resolve())
-    except assigned_actor.type.DoesNotExist:
-        return None
-
-
-def build_attachment_title(obj):
-    ev_metadata = obj.get_event_metadata()
-    ev_type = obj.get_event_type()
-
-    if ev_type == "error" and "type" in ev_metadata:
-        return ev_metadata["type"]
-    elif ev_type == "csp":
-        return "{} - {}".format(ev_metadata["directive"], ev_metadata["uri"])
-    else:
-        return obj.title
-
-
-def build_attachment_text(group, event=None):
-    # Group and Event both implement get_event_{type,metadata}
-    obj = event if event is not None else group
-    ev_metadata = obj.get_event_metadata()
-    ev_type = obj.get_event_type()
-
-    if ev_type == "error":
-        return ev_metadata.get("value") or ev_metadata.get("function")
-    else:
-        return None
-
-
-def build_assigned_text(group, identity, assignee):
-    actor = ActorTuple.from_actor_identifier(assignee)
-
-    try:
-        assigned_actor = actor.resolve()
-    except actor.type.DoesNotExist:
-        return
-
-    if actor.type == Team:
-        assignee_text = f"#{assigned_actor.slug}"
-    elif actor.type == User:
-        try:
-            assignee_ident = Identity.objects.get(
-                user=assigned_actor, idp__type="slack", idp__external_id=identity.idp.external_id
-            )
-            assignee_text = f"<@{assignee_ident.external_id}>"
-        except Identity.DoesNotExist:
-            assignee_text = assigned_actor.get_display_name()
-    else:
-        raise NotImplementedError
-
-    return f"*Issue assigned to {assignee_text} by <@{identity.external_id}>*"
-
-
-def build_action_text(group, identity, action):
-    if action["name"] == "assign":
-        return build_assigned_text(group, identity, action["selected_options"][0]["value"])
-
-    statuses = {"resolved": "resolved", "ignored": "ignored", "unresolved": "re-opened"}
-
-    # Resolve actions have additional 'parameters' after ':'
-    status = action["value"].split(":", 1)[0]
-
-    # Action has no valid action text, ignore
-    if status not in statuses:
-        return
-
-    return "*Issue {status} by <@{user_id}>*".format(
-        status=statuses[status], user_id=identity.external_id
-    )
-
-
-def build_rule_url(rule, group, project):
-    org_slug = group.organization.slug
-    project_slug = project.slug
-    rule_url = f"/organizations/{org_slug}/alerts/rules/{project_slug}/{rule.id}/"
-    return absolute_uri(rule_url)
-
-
-def build_group_attachment(
-    group, event=None, tags=None, identity=None, actions=None, rules=None, link_to_event=False
-):
-    # XXX(dcramer): options are limited to 100 choices, even when nested
-    status = group.get_status()
-
-    members = get_member_assignees(group)
-    teams = get_team_assignees(group)
-
-    logo_url = absolute_uri(get_asset_url("sentry", "images/sentry-email-avatar.png"))
-    text = build_attachment_text(group, event) or ""
-
-    if actions is None:
-        actions = []
-
-    assignee = get_assignee(group)
-
-    resolve_button = {
-        "name": "resolve_dialog",
-        "value": "resolve_dialog",
-        "type": "button",
-        "text": "Resolve...",
-    }
-
-    ignore_button = {"name": "status", "value": "ignored", "type": "button", "text": "Ignore"}
-
-    project = Project.objects.get_from_cache(id=group.project_id)
-
-    cache_key = "has_releases:2:%s" % (project.id)
-    has_releases = cache.get(cache_key)
-    if has_releases is None:
-        has_releases = ReleaseProject.objects.filter(project_id=project.id).exists()
-        if has_releases:
-            cache.set(cache_key, True, 3600)
-        else:
-            cache.set(cache_key, False, 60)
-
-    if not has_releases:
-        resolve_button.update({"name": "status", "text": "Resolve", "value": "resolved"})
-
-    if status == GroupStatus.RESOLVED:
-        resolve_button.update({"name": "status", "text": "Unresolve", "value": "unresolved"})
-
-    if status == GroupStatus.IGNORED:
-        ignore_button.update({"text": "Stop Ignoring", "value": "unresolved"})
-
-    option_groups = []
-
-    if teams:
-        option_groups.append({"text": "Teams", "options": teams})
-
-    if members:
-        option_groups.append({"text": "People", "options": members})
-
-    payload_actions = [
-        resolve_button,
-        ignore_button,
-        {
-            "name": "assign",
-            "text": "Select Assignee...",
-            "type": "select",
-            "selected_options": [assignee],
-            "option_groups": option_groups,
-        },
-    ]
-
-    # If an event is unspecified, use the tags of the latest event (if one exists).
-    event_for_tags = event if event else group.get_latest_event()
-
-    fallback_color = LEVEL_TO_COLOR["error"]
-    color = (
-        LEVEL_TO_COLOR.get(event_for_tags.get_tag("level"), fallback_color)
-        if event_for_tags
-        else fallback_color
-    )
-
-    fields = []
-    if tags:
-        event_tags = event_for_tags.tags if event_for_tags else []
-        for key, value in event_tags:
-            std_key = tagstore.get_standardized_key(key)
-            if std_key not in tags:
-                continue
-
-            labeled_value = tagstore.get_tag_value_label(key, value)
-            fields.append(
-                {
-                    "title": std_key.encode("utf-8"),
-                    "value": labeled_value.encode("utf-8"),
-                    "short": True,
-                }
-            )
-
-    if actions:
-        action_texts = [_f for _f in [build_action_text(group, identity, a) for a in actions] if _f]
-        text += "\n" + "\n".join(action_texts)
-
-        color = ACTIONED_ISSUE_COLOR
-        payload_actions = []
-
-    ts = group.last_seen
-
-    if event:
-        event_ts = event.datetime
-        ts = max(ts, event_ts)
-
-    footer = f"{group.qualified_short_id}"
-
-    if rules:
-        rule_url = build_rule_url(rules[0], group, project)
-        footer += f" via <{rule_url}|{rules[0].label}>"
-
-        if len(rules) > 1:
-            footer += f" (+{len(rules) - 1} other)"
-
-    obj = event if event is not None else group
-    if event and link_to_event:
-        title_link = group.get_absolute_url(params={"referrer": "slack"}, event_id=event.event_id)
-    else:
-        title_link = group.get_absolute_url(params={"referrer": "slack"})
-
-    return {
-        "fallback": f"[{project.slug}] {obj.title}",
-        "title": build_attachment_title(obj),
-        "title_link": title_link,
-        "text": text,
-        "fields": fields,
-        "mrkdwn_in": ["text"],
-        "callback_id": json.dumps({"issue": group.id}),
-        "footer_icon": logo_url,
-        "footer": footer,
-        "ts": to_timestamp(ts),
-        "color": color,
-        "actions": payload_actions,
-    }
-
-
-def build_incident_attachment(action, incident, metric_value=None, method=None):
-    """
-    Builds an incident attachment for slack unfurling
-    :param incident: The `Incident` to build the attachment for
-    :param metric_value: The value of the metric that triggered this alert to fire. If
-    not provided we'll attempt to calculate this ourselves.
-    :return:
-    """
-
-    data = incident_attachment_info(incident, metric_value, action=action, method=method)
-
-    colors = {
-        "Resolved": RESOLVED_COLOR,
-        "Warning": LEVEL_TO_COLOR["warning"],
-        "Critical": LEVEL_TO_COLOR["fatal"],
-    }
-
-    incident_footer_ts = (
-        "<!date^{:.0f}^Sentry Incident - Started {} at {} | Sentry Incident>".format(
-            to_timestamp(data["ts"]), "{date_pretty}", "{time}"
-        )
-    )
-
-    return {
-        "fallback": data["title"],
-        "title": data["title"],
-        "title_link": data["title_link"],
-        "text": data["text"],
-        "fields": [],
-        "mrkdwn_in": ["text"],
-        "footer_icon": data["logo_url"],
-        "footer": incident_footer_ts,
-        "color": colors[data["status"]],
-        "actions": [],
-    }
-
-
 # Different list types in slack that we'll use to resolve a channel name. Format is
 # (<list_name>, <result_name>, <prefix>).
-LIST_TYPES = [("conversations", "channels", CHANNEL_PREFIX), ("users", "members", MEMBER_PREFIX)]
+LIST_TYPES: List[Tuple[str, str, str]] = [
+    ("conversations", "channels", CHANNEL_PREFIX),
+    ("users", "members", MEMBER_PREFIX),
+]
 
 
-def strip_channel_name(name):
+def strip_channel_name(name: str):
     return name.lstrip(strip_channel_chars)
 
 
-def get_channel_id(organization, integration, name, use_async_lookup=False):
+def get_channel_id(
+    organization: Organization, integration: Integration, name: str, use_async_lookup: bool = False
+):
     """
     Fetches the internal slack id of a channel.
     :param organization: The organization that is using this integration
@@ -394,10 +96,12 @@ def validate_channel_id(name: str, integration_id: int, input_channel_id: str) -
         integration = Integration.objects.get(id=integration_id)
     except Integration.DoesNotExist:
         raise Http404
+
     token = integration.metadata["access_token"]
     headers = {"Authorization": f"Bearer {token}"}
     payload = {"channel": input_channel_id}
     client = SlackClient()
+
     try:
         results = client.get("/conversations.info", headers=headers, params=payload)
     except ApiError as e:
@@ -405,6 +109,10 @@ def validate_channel_id(name: str, integration_id: int, input_channel_id: str) -
             raise ValidationError("Channel not found. Invalid ID provided.")
         logger.info("rule.slack.conversation_info_failed", extra={"error": str(e)})
         raise IntegrationError("Could not retrieve Slack channel information.")
+
+    if not isinstance(results, dict):
+        raise IntegrationError("Bad slack channel list response.")
+
     stripped_channel_name = strip_channel_name(name)
     if not stripped_channel_name == results["channel"]["name"]:
         channel_name = results["channel"]["name"]
@@ -413,7 +121,7 @@ def validate_channel_id(name: str, integration_id: int, input_channel_id: str) -
         )
 
 
-def get_channel_id_with_timeout(integration, name, timeout):
+def get_channel_id_with_timeout(integration: Integration, name: str, timeout: int):
     """
     Fetches the internal slack id of a channel.
     :param integration: The slack integration
@@ -440,6 +148,8 @@ def get_channel_id_with_timeout(integration, name, timeout):
     client = SlackClient()
     id_data = None
     found_duplicate = False
+    prefix = ""
+
     for list_type, result_name, prefix in list_types:
         cursor = ""
         while True:
@@ -452,6 +162,9 @@ def get_channel_id_with_timeout(integration, name, timeout):
             except ApiError as e:
                 logger.info("rule.slack.%s_list_failed" % list_type, extra={"error": str(e)})
                 return (prefix, None, False)
+
+            if not isinstance(items, dict):
+                continue
 
             for c in items[result_name]:
                 # The "name" field is unique (this is the username for users)
@@ -484,6 +197,8 @@ def get_channel_id_with_timeout(integration, name, timeout):
 
 
 def send_incident_alert_notification(action, incident, metric_value, method):
+    from sentry.integrations.slack.message_builder.incidents import build_incident_attachment
+
     # Make sure organization integration is still active:
     try:
         integration = Integration.objects.get(
@@ -551,7 +266,6 @@ def parse_link(url):
         new_path.append(item)
 
     parsed_path = "/".join(new_path)
-
     parsed_path += "/" + str(url_parts[4])
 
     return parsed_path
